Remove existing config entry before adding new
* Moved `deinit_util` into `submodule_config_util`.  It's mostly
  concerned with the config file anyway, but I needed to use it from
  `SubmoduleConfigUtil` and didn't want to create a cycle.
* Factored out a method for removing entries from the config file.
* And wrote a couple of test cases for it -- it was basically untested
  before.
