+++ node/test/util/read_repo_ast_util.js
@@ -37,7 +37,6 @@ const NodeGit = require("nodegit");
 const path    = require("path");
 
 const ConflictUtil        = require("../../lib/util/conflict_util");
-const DeinitUtil          = require("../../lib/util/deinit_util");
 const Rebase              = require("../../lib/util/rebase");
 const RepoAST             = require("../../lib/util/repo_ast");
 const ReadRepoASTUtil     = require("../../lib/util/read_repo_ast_util");
@@ -544,7 +543,7 @@ describe("readRAST", function () {
 
         const commit = yield TestUtil.makeCommit(repo,
                                                  ["x/y", ".gitmodules"]);
-        yield DeinitUtil.deinit(repo, "x/y");
+        yield SubmoduleConfigUtil.deinit(repo, "x/y");
         let commits = {};
         commits[headCommit.id().tostrS()] = new Commit({
             changes: {"README.md":""},
@@ -592,7 +591,7 @@ describe("readRAST", function () {
         const index = yield repo.index();
         yield index.addByPath(".gitmodules");
         yield index.write();
-        yield DeinitUtil.deinit(repo, "x/y");
+        yield SubmoduleConfigUtil.deinit(repo, "x/y");
         let commits = {};
         commits[headCommit.id().tostrS()] = new Commit({
             changes: {"README.md":""},
@@ -643,7 +642,7 @@ describe("readRAST", function () {
         const subRepo = yield submodule.open();
         const anotherSubCommit = yield TestUtil.generateCommit(subRepo);
         const lastCommit = yield TestUtil.makeCommit(repo, ["x/y"]);
-        yield DeinitUtil.deinit(repo, "x/y");
+        yield SubmoduleConfigUtil.deinit(repo, "x/y");
 
         let commits = {};
         commits[headCommit.id().tostrS()] = new Commit({
@@ -804,7 +803,7 @@ describe("readRAST", function () {
                            baseSubPath,
                            "x/y",
                            subHead.id().tostrS());
-        yield DeinitUtil.deinit(repo, "x/y");
+        yield SubmoduleConfigUtil.deinit(repo, "x/y");
 
         let commits = {};
         commits[headCommit.id().tostrS()] = new Commit({
@@ -849,7 +848,7 @@ describe("readRAST", function () {
         const nextSubCommit = yield TestUtil.generateCommit(subRepo);
         const index = yield repo.index();
         yield index.addAll("x/y", -1);
-        yield DeinitUtil.deinit(repo, "x/y");
+        yield SubmoduleConfigUtil.deinit(repo, "x/y");
 
         let commits = {};
         commits[headCommit.id().tostrS()] = new Commit({
@@ -1527,7 +1526,7 @@ describe("readRAST", function () {
         const modules = ".gitmodules";
         const nextCommit = yield TestUtil.makeCommit(repo, ["a", modules]);
         const nextSha = nextCommit.id().tostrS();
-        yield DeinitUtil.deinit(repo, "a");
+        yield SubmoduleConfigUtil.deinit(repo, "a");
 
         yield fs.writeFile(path.join(repo.workdir(),
                                      modules),
